+++ README.md
@@ -20,7 +20,7 @@ import (
 
 func main() {
 	pfcpClient := client.New("1.2.3.4:8805")
-	recoveryTimeStamp := messages.NewRecoveryTimeStamp(time.Now())
+	recoveryTimeStamp := ie.NewRecoveryTimeStamp(time.Now())
 	sequenceNumber := uint32(21)
 
 	_, err := pfcpClient.SendHeartbeatRequest(recoveryTimeStamp, sequenceNumber)
