+++ main_test.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/dot-5g/pfcp/client"
-	"github.com/dot-5g/pfcp/information_elements"
+	"github.com/dot-5g/pfcp/ie"
 	"github.com/dot-5g/pfcp/messages"
 	"github.com/dot-5g/pfcp/server"
 )
@@ -14,14 +14,14 @@ import (
 var (
 	heartbeatRequestMu                        sync.Mutex
 	heartbeatRequesthandlerCalled             bool
-	heartbeatRequestreceivedRecoveryTimestamp information_elements.RecoveryTimeStamp
+	heartbeatRequestreceivedRecoveryTimestamp ie.RecoveryTimeStamp
 	heartbeatRequestReceivedSequenceNumber    uint32
 )
 
 var (
 	heartbeatResponseMu                        sync.Mutex
 	heartbeatResponsehandlerCalled             bool
-	heartbeatResponsereceivedRecoveryTimestamp information_elements.RecoveryTimeStamp
+	heartbeatResponsereceivedRecoveryTimestamp ie.RecoveryTimeStamp
 	heartbeatResponseReceivedSequenceNumber    uint32
 )
 
@@ -50,7 +50,7 @@ func HeartbeatRequest(t *testing.T) {
 	pfcpServer := server.New("127.0.0.1:8805")
 	pfcpServer.HeartbeatRequest(HandleHeartbeatRequest)
 	sentSequenceNumber := uint32(32)
-	recoveryTimeStamp := information_elements.NewRecoveryTimeStamp(time.Now())
+	recoveryTimeStamp := ie.NewRecoveryTimeStamp(time.Now())
 
 	go pfcpServer.Run()
 
@@ -82,7 +82,7 @@ func HeartbeatResponse(t *testing.T) {
 	pfcpServer := server.New("127.0.0.1:8805")
 	pfcpServer.HeartbeatResponse(HandleHeartbeatResponse)
 	sentSequenceNumber := uint32(971)
-	recoveryTimeStamp := information_elements.NewRecoveryTimeStamp(time.Now())
+	recoveryTimeStamp := ie.NewRecoveryTimeStamp(time.Now())
 
 	go pfcpServer.Run()
